[NoJira] refactor RN button component
